ChatAI-Back/requirements.txt
Flask==2.2.2
MarkupSafe==2.1.1
openai==0.25.0
pymongo==4.3.3
python-dotenv==0.21.1
